Move pagination control into separate file
