+++ src/components/controls/pagination-control/style.css
@@ -1,0 +1,14 @@
+.paginationControl {
+}
+
+.paginationControl input {
+    max-width: 5em;
+}
+
+.paginationControl > * {
+    margin: 0 0.15em;
+}
+
+.paginationControl label {
+    margin-right: 0.1em;
+}
