+++ src/components/entity-table/container/index.tsx
@@ -1,15 +1,14 @@
-import {useEffect} from "preact/hooks";
-
 import style from './style.css';
 
 import IcatClient from '../../../icat';
 import EntityTableView from '../view';
-import {randomSuffix, range} from '../../../utils';
+import {range} from '../../../utils';
 import {OpenRelatedHandler} from "../../context-menu";
 import {useQuery, useQueryClient} from "@tanstack/react-query";
 import OnChangeInput from "../../generic/on-change-input";
 import {EntityStateAction} from "../../../entityState";
 import {EntityTabState, NewIcatEntity, TableFilter} from "../../../types";
+import PaginationControl from "../../controls/pagination-control";
 
 type Props = {
     server: string;
@@ -115,66 +114,6 @@ const EntityTable = (
                 idx={idx}
             />}
     </>);
-}
-
-type PaginationProps = {
-    pageNumber: number;
-    handleSetPage: (n: number) => void;
-    handleLimitChange: (n: number) => void;
-    handlePageChange: (n: number) => void;
-}
-
-const PaginationControl = (
-    {
-        pageNumber, handleSetPage, handleLimitChange, handlePageChange
-    }: PaginationProps) => {
-    const decPage = () => handlePageChange(-1);
-    const incPage = () => handlePageChange(1);
-
-    const suffix = randomSuffix();
-    const prevId = `previousPageBtn_${suffix}`;
-    const nextId = `nextPageBtn_${suffix}`;
-
-    const focusOkForPageChange = () =>
-        document.activeElement === document.body
-        || document.activeElement === document.getElementById(prevId)
-        || document.activeElement === document.getElementById(nextId);
-
-    useEffect(() => {
-        const changePage = (ev: KeyboardEvent) => {
-            if (ev.key !== "ArrowLeft" && ev.key !== "ArrowRight") return;
-            if (!focusOkForPageChange()) return;
-
-            ev.preventDefault();
-            if (ev.key === "ArrowLeft") document.getElementById(prevId)!.click();
-            else document.getElementById(nextId)!.click();
-        };
-        document.addEventListener("keydown", changePage);
-        return () => {
-            document.removeEventListener("keydown", changePage);
-        }
-    });
-
-    return (
-        <span class={style.paginationControl}>
-            <button onClick={decPage} id={prevId}>Previous</button>
-            <input type="number"
-                   value={pageNumber}
-                   class={style.pageInput}
-                   onChange={ev => handleSetPage(
-                       parseInt((ev.target as HTMLInputElement).value))}/>
-            <button onClick={incPage} id={nextId}>Next</button>
-            <span>
-                <label for="pageSizeInput">Per page:</label>
-                <select name="pageSizeInput" onChange={
-                    ev => handleLimitChange(Number.parseInt((ev.target as HTMLSelectElement).value))}>
-                    <option value="20">20</option>
-                    <option value="50" selected>50</option>
-                    <option value="100">100</option>
-                </select>
-            </span>
-        </span>
-    );
 }
 
 type DeleteProps = {
